Rename: 팀 빌딩 컴포넌트 폴더 구조 변경, 레거시 파일 삭제
